Components ordered, in folders and separated, styles files added
