+++ AldhaNote/src/Views/NoteScreens/HomeScreen.tsx
@@ -6,14 +6,13 @@ import {
   ScrollView,
   RefreshControl,
   Text,
-  StyleSheet,
   Pressable,
 } from 'react-native';
 import {FAB} from 'react-native-paper';
-import CustomModal from '../components/Modal';
-import {useNotes} from '../hooks/useNotes';
-import ListItem from '../components/ListItem';
-import {noteDetailsRoute, editNoteRoute} from '../utils/route.utils';
+import CustomModal from '../../components/Modal';
+import {useNotes} from '../../hooks/useNotes';
+import ListItem from '../../components/ListItem';
+import {noteDetailsRoute, editNoteRoute} from '../../utils/route.utils';
 import {useIsFocused} from '@react-navigation/native';
 import {FontAwesomeIcon} from '@fortawesome/react-native-fontawesome';
 import {
@@ -21,6 +20,7 @@ import {
   faNoteSticky,
   faImage,
 } from '@fortawesome/free-solid-svg-icons';
+import {noteStyles} from '../styles';
 
 const wait = timeout => {
   return new Promise(resolve => setTimeout(resolve, timeout));
@@ -36,7 +36,6 @@ const HomeScreen = ({navigation}) => {
     getSpecificNotes,
     addNote,
     deleteNote,
-    editNote,
     setTrigger,
   } = useNotes();
   const [createModalVisible, setCreateModalVisible] = useState(false);
@@ -75,62 +74,70 @@ const HomeScreen = ({navigation}) => {
   };
 
   return (
-    <SafeAreaView style={styles.bg}>
+    <SafeAreaView style={noteStyles.bg}>
       {
-        <View style={styles.container}>
+        <View style={noteStyles.container}>
           <ScrollView
             horizontal={true}
-            contentContainerStyle={styles.scrollViewHorizontal}>
+            contentContainerStyle={noteStyles.scrollViewHorizontalHome}>
             <Pressable
               style={
-                currentType === 'all' ? styles.selected : styles.unselected
+                currentType === 'all'
+                  ? noteStyles.selected
+                  : noteStyles.unselected
               }
               onPress={() => {
                 setCurrentType('all');
                 setTrigger(!trigger);
               }}>
               <FontAwesomeIcon size={20} icon={faPenToSquare} color={'white'} />
-              <Text style={styles.option}>All</Text>
+              <Text style={noteStyles.option}>All</Text>
             </Pressable>
             <Pressable
               style={
-                currentType === 'note' ? styles.selected : styles.unselected
+                currentType === 'note'
+                  ? noteStyles.selected
+                  : noteStyles.unselected
               }
               onPress={() => {
                 setCurrentType('note');
                 setTrigger(!trigger);
               }}>
               <FontAwesomeIcon size={20} icon={faNoteSticky} color={'white'} />
-              <Text style={styles.option}>Note</Text>
+              <Text style={noteStyles.option}>Note</Text>
             </Pressable>
             <Pressable
               style={
-                currentType === 'image' ? styles.selected : styles.unselected
+                currentType === 'image'
+                  ? noteStyles.selected
+                  : noteStyles.unselected
               }
               onPress={() => {
                 setCurrentType('image');
                 setTrigger(!trigger);
               }}>
               <FontAwesomeIcon size={20} icon={faImage} color={'white'} />
-              <Text style={styles.option}>Image</Text>
+              <Text style={noteStyles.option}>Image</Text>
             </Pressable>
             <Pressable
               style={
-                currentType === 'list' ? styles.selected : styles.unselected
+                currentType === 'list'
+                  ? noteStyles.selected
+                  : noteStyles.unselected
               }
               onPress={() => {
                 setCurrentType('list');
                 setTrigger(!trigger);
               }}>
               <FontAwesomeIcon size={20} icon={faPenToSquare} color={'white'} />
-              <Text style={styles.option}>List</Text>
+              <Text style={noteStyles.option}>List</Text>
             </Pressable>
           </ScrollView>
         </View>
       }
       {isLoading && <ActivityIndicator size="large" />}
       <ScrollView
-        contentContainerStyle={styles.scrollView}
+        contentContainerStyle={noteStyles.scrollViewHome}
         refreshControl={
           <RefreshControl refreshing={refreshing} onRefresh={onRefresh} />
         }>
@@ -150,7 +157,7 @@ const HomeScreen = ({navigation}) => {
             />
           ))}
         {!isLoading && !hasElements && (
-          <Text style={styles.title}>No notes yet :C</Text>
+          <Text style={noteStyles.title}>No notes yet :C</Text>
         )}
 
         <CustomModal
@@ -160,7 +167,7 @@ const HomeScreen = ({navigation}) => {
         />
       </ScrollView>
       <FAB
-        style={styles.fab}
+        style={noteStyles.fab}
         small
         icon="plus"
         onPress={() => setCreateModalVisible(true)}
@@ -168,75 +175,5 @@ const HomeScreen = ({navigation}) => {
     </SafeAreaView>
   );
 };
-
-const styles = StyleSheet.create({
-  view: {
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'center',
-    minHeight: '100%',
-  },
-  fab: {
-    position: 'absolute',
-    margin: 16,
-    right: 0,
-    bottom: 0,
-  },
-  title: {
-    marginRight: 16,
-    marginLeft: 16,
-    marginTop: 16,
-    fontSize: 20,
-    fontWeight: 'bold',
-  },
-  item: {
-    backgroundColor: '#f9c2ff',
-    padding: 20,
-    marginVertical: 8,
-    marginHorizontal: 16,
-  },
-  bg: {
-    flex: 1,
-    position: 'relative',
-    marginTop: 20,
-  },
-  container: {
-    alignItems: 'center',
-    marginBottom: 10,
-  },
-  scrollView: {
-    alignItems: 'center',
-  },
-  scrollViewHorizontal: {
-    alignItems: 'flex-start',
-  },
-  option: {
-    marginLeft: 5,
-    fontSize: 15,
-    color: 'white',
-    textAlign: 'center',
-  },
-  unselected: {
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-    alignItems: 'center',
-    backgroundColor: '#A6A6A6',
-    margin: 5,
-    padding: 10,
-    borderRadius: 10,
-  },
-  selected: {
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-    alignItems: 'center',
-    backgroundColor: '#575757',
-    margin: 5,
-    padding: 10,
-    borderRadius: 10,
-  },
-});
 
 export default HomeScreen;
