+++ AldhaNote/src/components/Button/styles.tsx
@@ -1,0 +1,29 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  button: {
+    width: '90%',
+    height: 60,
+    marginVertical: 10,
+    alignItems: 'center',
+    backgroundColor: '#2e2e2e',
+    justifyContent: 'center',
+    borderRadius: 10,
+  },
+  buttonSeparated: {
+    width: '40%',
+  },
+  buttonText: {
+    color: 'white',
+    fontSize: 18,
+  },
+  disabled: {
+    width: '90%',
+    height: 60,
+    marginVertical: 10,
+    alignItems: 'center',
+    backgroundColor: '#949494',
+    justifyContent: 'center',
+    borderRadius: 10,
+  },
+});
