+++ AldhaNote/src/components/CustomInput/index.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
-import {StyleSheet, View, Text} from 'react-native';
+import {View, Text} from 'react-native';
 import {TextInput} from 'react-native-paper';
+import {styles} from './styles';
 
-const Input = ({
+const CustomInput = ({
   label,
   placeholder,
   value,
@@ -13,15 +14,26 @@ const Input = ({
   onBlur,
   error,
   onRightIconPress,
+  isMultiline = false,
+  numLines = 1,
+  id = -1,
+  isScreen = false,
 }) => {
   return (
-    <View style={styles.view}>
+    <View
+      style={!isScreen ? styles.view : {...styles.view, ...styles.fullScreen}}>
       <TextInput
         label={label}
         placeholder={placeholder}
         value={value}
+        multiline={isMultiline}
+        numberOfLines={isMultiline ? numLines : 1}
         secureTextEntry={isSecured}
-        onChangeText={onChangeText}
+        onChangeText={text =>
+          id === -1
+            ? onChangeText(placeholder, text)
+            : onChangeText(id, text, id)
+        }
         onBlur={onBlur}
         style={styles.input}
         left={!!leftIcon && <TextInput.Icon name={leftIcon} />}
@@ -36,21 +48,4 @@ const Input = ({
   );
 };
 
-const styles = StyleSheet.create({
-  view: {
-    position: 'relative',
-  },
-  input: {
-    marginBottom: 40,
-  },
-  error: {
-    position: 'absolute',
-    color: 'red',
-    bottom: 10,
-    textAlign: 'center',
-    fontSize: 10,
-    marginLeft: 10,
-  },
-});
-
-export default Input;
+export default CustomInput;
