+++ AldhaNote/src/components/CustomSelect/styles.tsx
@@ -1,0 +1,25 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  text: {},
+  touchableOpacity: {
+    flex: 1,
+    backgroundColor: '#fff',
+    borderRadius: 10,
+    padding: 10,
+    marginTop: 2,
+    marginHorizontal: 4,
+    marginBottom: 30,
+    elevation: 2,
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 4,
+  },
+  touchableContainer: {
+    height: 80,
+  },
+});
