+++ AldhaNote/src/components/Input/styles.tsx
@@ -1,0 +1,18 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  view: {
+    position: 'relative',
+  },
+  input: {
+    marginBottom: 40,
+  },
+  error: {
+    position: 'absolute',
+    color: 'red',
+    bottom: 10,
+    textAlign: 'center',
+    fontSize: 10,
+    marginLeft: 10,
+  },
+});
