+++ AldhaNote/src/components/ListItem/index.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import {View, Text, TouchableOpacity, StyleSheet} from 'react-native';
-import {windowWidth} from '../utils/dimensions.utils';
+import {View, Text, TouchableOpacity} from 'react-native';
 import {FontAwesomeIcon} from '@fortawesome/react-native-fontawesome';
 import {
   faPenToSquare,
@@ -8,6 +7,8 @@ import {
   faNoteSticky,
   faImage,
 } from '@fortawesome/free-solid-svg-icons';
+import {windowWidth} from '../../utils/dimensions.utils';
+import {styles} from './styles';
 
 export default function ListItem({
   icon,
@@ -56,49 +57,3 @@ export default function ListItem({
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  listIcon: {
-    marginRight: 15,
-  },
-  listItem: {
-    backgroundColor: '#FFF',
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: 20,
-    width: windowWidth - 30,
-    padding: 10,
-    borderRadius: 10,
-    shadowColor: '#171717',
-    elevation: 10,
-  },
-  listView: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    flex: 1,
-    marginLeft: 10,
-  },
-  listSubTitle: {
-    color: '#333',
-    fontFamily: 'Roboto-Medium',
-    fontSize: 14,
-  },
-  listTitle: {
-    color: '#333',
-    fontFamily: 'Roboto-Medium',
-    fontSize: 14,
-    textTransform: 'uppercase',
-  },
-  touchable: {
-    width: 20,
-    borderRadius: 10,
-    marginRight: 10,
-  },
-  text: {
-    color: '#fff',
-    textAlign: 'center',
-    fontFamily: 'Roboto-Medium',
-    fontSize: 14,
-  },
-});
