+++ AldhaNote/src/components/Login/styles.tsx
@@ -1,0 +1,18 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {
+    marginRight: 12,
+    marginLeft: 12,
+    height: '100%',
+  },
+  title: {
+    textAlign: 'center',
+    fontSize: 20,
+    marginTop: 20,
+  },
+  backgroundStyle: {},
+  button: {
+    marginBottom: 16,
+  },
+});
