+++ AldhaNote/src/components/Register/styles.tsx
@@ -1,0 +1,33 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {
+    marginRight: 12,
+    marginLeft: 12,
+    height: '100%',
+  },
+  title: {
+    textAlign: 'center',
+    fontSize: 20,
+    marginTop: 20,
+  },
+  backgroundStyle: {},
+  button: {
+    marginBottom: 16,
+  },
+  image: {
+    borderRadius: 10,
+    minWidth: 150,
+    minHeight: 150,
+  },
+  imageContainer: {
+    borderRadius: 10,
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    alignItems: 'center',
+    borderColor: '#000',
+    borderWidth: 2,
+    padding: 10,
+  },
+});
